requirements.txt
streamlit
streamlit-autorefresh
Pillow
pyzbar
requests
pyotp
boto3